requirements.txt
numpy>=1.24
gymnasium>=0.29.1
stable-baselines3>=2.3.0
torch>=2.1.0
matplotlib>=3.8
pyyaml>=6.0.1
tqdm>=4.66